IPython Console: Create a new module called jupyter_qtconsole and move there several related widgets
