+++ spyder/widgets/jupyter_qtconsole/__init__.py
@@ -1,0 +1,13 @@
+# -*- coding:utf-8 -*-
+#
+# Copyright © Spyder Project Contributors
+# Licensed under the terms of the MIT License
+# (see spyder/__init__.py for details)
+
+"""
+Widgets for our implementation of QtConsole
+"""
+
+from .control import ControlWidget, PageControlWidget
+from .help import HelpWidget
+from .namespacebrowser import NamepaceBrowserWidget
